Add files via upload
This has nothing to do with solving the challenge. When I was trying to figure out what the challenge was about I drew a bunch of lines connecting similar segments to try to figure out if the lines might reveal a pattern or something. When I was done adding the lines I realized it looked like I was trying to prove a conspiracy theory.